refactor: Move fixture into conftest
